Move methods to ObjectUnderFileSystem
