+++ core/common/src/main/java/alluxio/underfs/ObjectUnderFileSystem.java
@@ -12,6 +12,17 @@
 package alluxio.underfs;
 
 import alluxio.AlluxioURI;
+import alluxio.Configuration;
+import alluxio.Constants;
+import alluxio.PropertyKey;
+import alluxio.underfs.options.CreateOptions;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.IOException;
+import java.io.OutputStream;
+import java.util.List;
 
 import javax.annotation.concurrent.ThreadSafe;
 
@@ -20,6 +31,8 @@ import javax.annotation.concurrent.ThreadSafe;
  */
 @ThreadSafe
 public abstract class ObjectUnderFileSystem extends BaseUnderFileSystem {
+  private static final Logger LOG = LoggerFactory.getLogger(Constants.LOGGER_TYPE);
+
   /**
    * Constructs an {@link ObjectUnderFileSystem}.
    *
@@ -28,6 +41,68 @@ public abstract class ObjectUnderFileSystem extends BaseUnderFileSystem {
   protected ObjectUnderFileSystem(AlluxioURI uri) {
     super(uri);
   }
+
+  @Override
+  public void close() throws IOException {
+  }
+
+  @Override
+  public void connectFromMaster(String hostname) {
+    // Authentication is taken care of in the constructor
+  }
+
+  @Override
+  public void connectFromWorker(String hostname) {
+    // Authentication is taken care of in the constructor
+  }
+
+  @Override
+  public OutputStream create(String path, CreateOptions options) throws IOException {
+    return createDirect(path, options);
+  }
+
+  /**
+   * Gets the block size in bytes. This method defaults to the default user block size in Alluxio.
+   *
+   * @param path the file name
+   * @return the default Alluxio user block size
+   * @throws IOException this implementation will not throw this exception, but subclasses may
+   */
+  @Override
+  public long getBlockSizeByte(String path) throws IOException {
+    return Configuration.getBytes(PropertyKey.USER_BLOCK_SIZE_BYTES_DEFAULT);
+  }
+
+  // Not supported
+  @Override
+  public Object getConf() {
+    LOG.debug("getConf is not supported when using default ObjectUnderFileSystem.");
+    return null;
+  }
+
+  // Not supported
+  @Override
+  public List<String> getFileLocations(String path) throws IOException {
+    LOG.debug("getFileLocations is not supported when using default ObjectUnderFileSystem.");
+    return null;
+  }
+
+  // Not supported
+  @Override
+  public List<String> getFileLocations(String path, long offset) throws IOException {
+    LOG.debug("getFileLocations is not supported when using default ObjectUnderFileSystem.");
+    return null;
+  }
+
+  // This call is currently only used for the web ui, where a negative value implies unknown.
+  @Override
+  public long getSpace(String path, SpaceType type) throws IOException {
+    return -1;
+  }
+
+  // Default object UFS does not provide a mechanism for updating the configuration, no-op
+  @Override
+  public void setConf(Object conf) {}
 
   @Override
   public boolean supportsFlush() {
