+++ underfs/gcs/src/main/java/alluxio/underfs/gcs/GCSUnderFileSystem.java
@@ -44,7 +44,6 @@ import java.io.OutputStream;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import javax.annotation.concurrent.ThreadSafe;
@@ -157,25 +156,6 @@ public final class GCSUnderFileSystem extends ObjectUnderFileSystem {
   }
 
   @Override
-  public void close() throws IOException {
-  }
-
-  @Override
-  public void connectFromMaster(String hostname) {
-    // Authentication is taken care of in the constructor
-  }
-
-  @Override
-  public void connectFromWorker(String hostname) {
-    // Authentication is taken care of in the constructor
-  }
-
-  @Override
-  public OutputStream create(String path, CreateOptions options) throws IOException {
-    return createDirect(path, options);
-  }
-
-  @Override
   public OutputStream createDirect(String path, CreateOptions options) throws IOException {
     if (mkdirs(getParentKey(path), true)) {
       return new GCSOutputStream(mBucketName, stripPrefixIfPresent(path), mClient);
@@ -227,40 +207,6 @@ public final class GCSUnderFileSystem extends ObjectUnderFileSystem {
     return deleteInternal(path);
   }
 
-  /**
-   * Gets the block size in bytes. There is no concept of a block in GCS and the maximum size of
-   * one file is 5 TB. This method defaults to the default user block size in Alluxio.
-   *
-   * @param path the file name
-   * @return the default Alluxio user block size
-   * @throws IOException this implementation will not throw this exception, but subclasses may
-   */
-  @Override
-  public long getBlockSizeByte(String path) throws IOException {
-    return Configuration.getBytes(PropertyKey.USER_BLOCK_SIZE_BYTES_DEFAULT);
-  }
-
-  // Not supported
-  @Override
-  public Object getConf() {
-    LOG.debug("getConf is not supported when using GCSUnderFileSystem, returning null.");
-    return null;
-  }
-
-  // Not supported
-  @Override
-  public List<String> getFileLocations(String path) throws IOException {
-    LOG.debug("getFileLocations is not supported when using GCSUnderFileSystem, returning null.");
-    return null;
-  }
-
-  // Not supported
-  @Override
-  public List<String> getFileLocations(String path, long offset) throws IOException {
-    LOG.debug("getFileLocations is not supported when using GCSUnderFileSystem, returning null.");
-    return null;
-  }
-
   @Override
   public long getFileSize(String path) throws IOException {
     GSObject details = getObjectDetails(path);
@@ -279,12 +225,6 @@ public final class GCSUnderFileSystem extends ObjectUnderFileSystem {
     } else {
       throw new FileNotFoundException(path);
     }
-  }
-
-  // This call is currently only used for the web ui, where a negative value implies unknown.
-  @Override
-  public long getSpace(String path, SpaceType type) throws IOException {
-    return -1;
   }
 
   @Override
@@ -452,10 +392,6 @@ public final class GCSUnderFileSystem extends ObjectUnderFileSystem {
     // Source is a file and Destination does not exist
     return copy(src, dst) && deleteInternal(src);
   }
-
-  // Not supported
-  @Override
-  public void setConf(Object conf) {}
 
   // Setting GCS owner via Alluxio is not supported yet. This is a no-op.
   @Override
