+++ underfs/s3a/src/main/java/alluxio/underfs/s3a/S3AUnderFileSystem.java
@@ -55,7 +55,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 import javax.annotation.concurrent.ThreadSafe;
@@ -225,25 +224,6 @@ public class S3AUnderFileSystem extends ObjectUnderFileSystem {
   }
 
   @Override
-  public void close() throws IOException {
-  }
-
-  @Override
-  public void connectFromMaster(String hostname) {
-    // Authentication is taken care of in the constructor
-  }
-
-  @Override
-  public void connectFromWorker(String hostname) {
-    // Authentication is taken care of in the constructor
-  }
-
-  @Override
-  public OutputStream create(String path, CreateOptions options) throws IOException {
-    return createDirect(path, options);
-  }
-
-  @Override
   public OutputStream createDirect(String path, CreateOptions options) throws IOException {
     if (mkdirs(getParentKey(path), true)) {
       return new S3AOutputStream(mBucketName, stripPrefixIfPresent(path), mManager);
@@ -289,41 +269,6 @@ public class S3AUnderFileSystem extends ObjectUnderFileSystem {
     return deleteInternal(path);
   }
 
-  /**
-   * Gets the block size in bytes. There is no concept of a block in S3 and the maximum size of
-   * one put is 5 GB, and the max size of a multi-part upload is 5 TB. This method defaults to the
-   * default user block size in Alluxio.
-   *
-   * @param path the file name
-   * @return the default Alluxio user block size
-   * @throws IOException this implementation will not throw this exception, but subclasses may
-   */
-  @Override
-  public long getBlockSizeByte(String path) throws IOException {
-    return Configuration.getBytes(PropertyKey.USER_BLOCK_SIZE_BYTES_DEFAULT);
-  }
-
-  // Not supported
-  @Override
-  public Object getConf() {
-    LOG.debug("getConf is not supported when using S3AUnderFileSystem, returning null.");
-    return null;
-  }
-
-  // Not supported
-  @Override
-  public List<String> getFileLocations(String path) throws IOException {
-    LOG.debug("getFileLocations is not supported when using S3AUnderFileSystem, returning null.");
-    return null;
-  }
-
-  // Not supported
-  @Override
-  public List<String> getFileLocations(String path, long offset) throws IOException {
-    LOG.debug("getFileLocations is not supported when using S3AUnderFileSystem, returning null.");
-    return null;
-  }
-
   @Override
   public long getFileSize(String path) throws IOException {
     try {
@@ -343,12 +288,6 @@ public class S3AUnderFileSystem extends ObjectUnderFileSystem {
     } else {
       throw new FileNotFoundException(path);
     }
-  }
-
-  // This call is currently only used for the web ui, where a negative value implies unknown.
-  @Override
-  public long getSpace(String path, SpaceType type) throws IOException {
-    return -1;
   }
 
   @Override
@@ -495,10 +434,6 @@ public class S3AUnderFileSystem extends ObjectUnderFileSystem {
     // Source is a file and Destination does not exist
     return copy(src, dst) && deleteInternal(src);
   }
-
-  // This UFS does not provide a mechanism for updating the configuration, no-op
-  @Override
-  public void setConf(Object conf) {}
 
   // Setting S3 owner via Alluxio is not supported yet. This is a no-op.
   @Override
